docs(nix): update theme name capitalization again (#122)
This reverts commit eb7ed7ab2ca97730fcc5842316a132e60f7446ba.

f13a2b2 changed the naming convention for the theme names again so that
the theme style (dark or light) is capitalized again. The docs were
updated accordingly for Home Manager.

However, eb7ed7a changed the documentation for Home Manager back to the
non-capitalized names. This was likely because the capitalization change
had not made its way to `nixos-unstable`. The change has now made it to
`nixos-unstable`, so this commit reverts eb7ed7 to reflect the
capitalization.